backend/requirements.txt
# Core Django
Django>=4.2,<5.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.0

# CORS handling for Flutter frontend
django-cors-headers>=4.0.0

# Environment variables
python-decouple>=3.8

# Production server
gunicorn>=21.2.0

# Development tools
django-debug-toolbar>=4.0.0  # Development only
pytest>=7.4.0
pytest-django>=4.5.0

# Code quality
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0
